<NLBackend> #22 feat: update ledger total income and expense when create transaction

+++ NestLedger/src/main/java/org/xxooooxx/nestledger/vo/ledger/request/LedgerUpdateRequestData.java
@@ -22,7 +22,7 @@ public class LedgerUpdateRequestData {
 
     private String title;
     private List<String> userIds;
-    private BigDecimal totalIncome;
-    private BigDecimal totalExpense;
+    private Integer totalIncome;
+    private Integer totalExpense;
     private Integer version;
 }
