+++ NestLedger/src/main/java/org/xxooooxx/nestledger/vo/ledger/response/LedgerGetResponseData.java
@@ -20,8 +20,8 @@ public class LedgerGetResponseData {
     private String _id;
     private String title;
     private ArrayList<String> userIds;
-    private BigDecimal totalIncome;
-    private BigDecimal totalExpense;
+    private Integer totalIncome;
+    private Integer totalExpense;
     private Integer version;
 
     public LedgerGetResponseData(LedgerDB data) {
